fix gtm implementation
