+++ src/frontend/middleware/send-analytics.ts
@@ -2,7 +2,7 @@ import {Injectable} from '@angular/core';
 import {MainActions} from '../actions/main-actions';
 import {Options} from '../state';
 import {AnalyticsConsent} from '../../options';
-import GoogleTagManagerSend from '../../gtm';
+import {MessageFactory} from '../../communication';
 
 @Injectable()
 export class SendAnalytics {
@@ -11,10 +11,9 @@ export class SendAnalytics {
   middleware = store => next => action => {
     if (action.type === MainActions.SEND_ANALYTICS &&
       this.options.analyticsConsent === AnalyticsConsent.Yes) {
-      GoogleTagManagerSend({
-        'event': action.payload.event,
-        'desc': action.payload.desc
-      });
+      chrome.runtime.sendMessage(MessageFactory.analyticsEvent(
+        action.payload.event,
+        action.payload.desc));
     }
     return next(action);
   }
